# openapi-v1.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/notifications/{id}/read": {
      "put": {
        "operationId": "NotificationsV1Controller_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "v1-notifications"
        ]
      }
    },
...
    "/v1/messages": {
      "get": {
        "operationId": "MessagesV1Controller_conversations",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor returned by the previous page.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor-paginated direct message conversations."
          }
        },
        "tags": [
          "v1-messages"
        ]
      },
      "post": {
        "operationId": "MessagesV1Controller_send",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageV1Dto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message sent using the existing block and notification policy."
          }
        },
        "tags": [
          "v1-messages"
        ]
      }
    },
...
    "/v1/messages/{userId}": {
      "get": {
        "operationId": "MessagesV1Controller_conversation",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor returned by the previous page.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cursor-paginated direct conversation. Reading marks incoming messages as read."
          }
        },
        "tags": [
          "v1-messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "MDTBBS Public Client API",
    "description": "Stable V1 contract for web, official applications and third-party public clients. Use /api/v1/capabilities for feature discovery. External server integrations use the separate /api/external/v1 API key boundary. JSON V1 endpoints return the { data, meta } / { error, meta } envelope unless explicitly documented as a raw file or image response.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "Same-origin MindFourm API"
    }
  ],
  "components": {
    "securitySchemes": {
      "MindAuthBearer": {
        "scheme": "bearer",
        "bearerFormat": "MindAuth access token",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMessageV1Dto": {
        "type": "object",
        "properties": {
          "recipient_id": {
            "type": "number",
            "minimum": 1
          },
          "content": {
            "type": "string",
            "maxLength": 5000
          }
        },
        "required": [
          "recipient_id",
          "content"
        ]
      },
...
    }
  }
}